Now checks to see if there are .cpp files first before generating makefile

+++ MakeGen.sh
@@ -16,22 +16,25 @@ if [ ! -d $1 ] || [ ! $1 ] ; then
 # if they did pass us a directory then list all of the files and find the .cpp's 
 else
     cppFiles=$(ls -a $1 | grep .cpp)
+    if [ ! -z "$cppFiles" ] ; then
+	# this command searches all of the .cpp files for user created include statements (#include "Book.h")
+	# and formats it into an exceptable format for passing into the makeGen c++ program for further parsing and manipulation
+	includes=$(grep -i "#include" $1/*.cpp | sed 's/.*>//g' | sed 's/#include//g' | sed 's/.*\///g' | sed 's/\"//g' | sed 's/: /:/g')
 
-    # this command searches all of the .cpp files for user created include statements (#include "Book.h")
-    # and formats it into an exceptable format for passing into the makeGen c++ program for further parsing and manipulation
-    includes=$(grep -i "#include" $1/*.cpp | sed 's/.*>//g' | sed 's/#include//g' | sed 's/.*\///g' | sed 's/\"//g' | sed 's/: /:/g')
+	# runs the makeGen.cpp program with the cppFiles, include statements and directory argument
+	./makeG $cppFiles $includes $1
 
-    # runs the makeGen.cpp program with the cppFiles, include statements and directory argument
-    ./makeG $cppFiles $includes $1
+	# clean up the cpp program
+	rm makeG
 
-    # clean up the cpp program
-    rm makeG
+	# move the makefile into the given directory
+	mv makefile $1/
 
-    # move the makefile into the given directory
-    mv makefile $1/
-
-    # run the make command and back out
-    cd $1
-    make
-    cd $(pwd)
+	# run the make command and back out
+	cd $1
+	make
+	cd $(pwd)
+    else
+	echo "No .cpp files found in directory"
+    fi
 fi
